+++ elide-async/pom.xml
@@ -90,13 +90,6 @@
             <version>2.4.0</version>
         </dependency>
 
-        <dependency>
-            <groupId>javax.servlet</groupId>
-            <artifactId>javax.servlet-api</artifactId>
-            <version>4.0.1</version>
-        </dependency>
-
-
         <!-- Test -->
         <!-- JUnit -->
         <dependency>
@@ -147,12 +140,6 @@
             <groupId>javax.el</groupId>
             <artifactId>javax.el-api</artifactId>
             <version>3.0.0</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.glassfish.jersey.core</groupId>
-            <artifactId>jersey-server</artifactId>
-            <version>2.33</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
