+++ elide-standalone/pom.xml
@@ -200,6 +200,12 @@
             <scope>test</scope>
         </dependency>
 
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-core</artifactId>
+            <scope>test</scope>
+        </dependency>
+
         <!-- Elide Test Helpers -->
         <dependency>
             <groupId>com.yahoo.elide</groupId>
