+++ elide-standalone/src/main/java/com/yahoo/elide/standalone/resources/ExportApiEndpoint.java
@@ -3,7 +3,7 @@
  * Licensed under the Apache License, Version 2.0
  * See LICENSE file in project root for terms.
  */
-package com.yahoo.elide.async.resources;
+package com.yahoo.elide.standalone.resources;
 
 import com.yahoo.elide.async.service.storageengine.ResultStorageEngine;
 import com.yahoo.elide.core.exceptions.HttpStatus;
