+++ elide-standalone/src/test/java/example/ElideStandaloneDisableAggStoreTest.java
@@ -17,15 +17,10 @@ import static io.restassured.RestAssured.when;
 import static org.hamcrest.Matchers.containsInAnyOrder;
 import com.yahoo.elide.standalone.ElideStandalone;
 import com.yahoo.elide.standalone.config.ElideStandaloneAnalyticSettings;
-import com.yahoo.elide.standalone.config.ElideStandaloneAsyncSettings;
-import com.yahoo.elide.standalone.config.ElideStandaloneSettings;
-import example.models.Post;
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
-
-import java.util.Properties;
 
 /**
  * Tests ElideStandalone starts and works.
@@ -35,75 +30,7 @@ public class ElideStandaloneDisableAggStoreTest extends ElideStandaloneTest {
 
     @BeforeAll
     public void init() throws Exception {
-        elide = new ElideStandalone(new ElideStandaloneSettings() {
-
-            @Override
-            public Properties getDatabaseProperties() {
-                Properties options = new Properties();
-
-                options.put("hibernate.show_sql", "true");
-                options.put("hibernate.hbm2ddl.auto", "create");
-                options.put("hibernate.dialect", "org.hibernate.dialect.H2Dialect");
-                options.put("hibernate.current_session_context_class", "thread");
-                options.put("hibernate.jdbc.use_scrollable_resultset", "true");
-
-                options.put("javax.persistence.jdbc.driver", "org.h2.Driver");
-                options.put("javax.persistence.jdbc.url", "jdbc:h2:mem:db1;DB_CLOSE_DELAY=-1;");
-                options.put("javax.persistence.jdbc.user", "sa");
-                options.put("javax.persistence.jdbc.password", "");
-                return options;
-            }
-
-            @Override
-            public String getModelPackageName() {
-                return Post.class.getPackage().getName();
-            }
-
-            @Override
-            public boolean enableSwagger() {
-                return true;
-            }
-
-            @Override
-            public boolean enableGraphQL() {
-                return true;
-            }
-
-            @Override
-            public boolean enableJSONAPI() {
-                return true;
-            }
-
-            @Override
-            public ElideStandaloneAsyncSettings getAsyncProperties() {
-                ElideStandaloneAsyncSettings asyncPropeties = new ElideStandaloneAsyncSettings() {
-                    @Override
-                    public boolean enabled() {
-                        return true;
-                    }
-
-                    @Override
-                    public boolean enableCleanup() {
-                        return true;
-                    }
-
-                    @Override
-                    public Integer getThreadSize() {
-                        return 5;
-                    }
-
-                    @Override
-                    public Integer getMaxRunTimeSeconds() {
-                        return 1800;
-                    }
-
-                    @Override
-                    public Integer getQueryCleanupDays() {
-                        return 3;
-                    }
-                };
-                return asyncPropeties;
-            }
+        elide = new ElideStandalone(new ElideStandaloneTestSettings() {
 
             @Override
             public ElideStandaloneAnalyticSettings getAnalyticProperties() {
